+++ src/server/OTERRAIN.cpp
@@ -296,7 +296,7 @@ void TerrainRes::load_info()
 	TerrainRec  	*terrainRec;
 	TerrainInfo 	*terrainInfo = NULL;
 	int      	 	i;
-	long			 	bitmapOffset;
+	uint32_t			bitmapOffset;
 
 	//---- read in terrain count and initialize terrain info array ----//
 
@@ -412,7 +412,7 @@ void TerrainRes::load_info()
 
 		//---- get the bitmap pointer of the terrain icon in res_icon ----//
 
-		memcpy( &bitmapOffset, terrainRec->bitmap_ptr, sizeof(long) );
+		memcpy( &bitmapOffset, terrainRec->bitmap_ptr, sizeof(uint32_t) );
 
 		terrainInfo->bitmap_ptr = res_bitmap.read_imported(bitmapOffset);
 
@@ -784,7 +784,7 @@ void TerrainRes::load_anim_info()
 {
 	TerrainAnimRec	terrainAnimRec, lastAnimRec;
 	int i;
-	long bitmapOffset;
+	uint32_t bitmapOffset;
 
 	String terAnimDbName(DIR_RES);
 	terAnimDbName += "TERANM";
@@ -808,7 +808,7 @@ void TerrainRes::load_anim_info()
 		char* bitmapPtr;
 		terrainAnimRec = *(TerrainAnimRec *)(dbTerAnim->read(i+1));
 		
-		memcpy( &bitmapOffset, terrainAnimRec.bitmap_ptr, sizeof(long) );
+		memcpy( &bitmapOffset, terrainAnimRec.bitmap_ptr, sizeof(uint32_t) );
 		bitmapPtr = anm_bitmap.read_imported(bitmapOffset);
 
 		if( memcmp(terrainAnimRec.base_file, lastAnimRec.base_file, terrainAnimRec.FILE_NAME_LEN))
